Add synchronization to category repository

+++ data/src/main/java/com/example/data/repositories/CategoryRepositoryImpl.kt
@@ -20,17 +20,38 @@ class CategoryRepositoryImpl @Inject constructor(
 
     override suspend fun createCategory(category: Category) {
         val local = mapper.map(category)
-        localSource.addCategory(local)
+        val id = localSource.addCategory(local)
+
+        if (!tokenManager.isAuthorized()) return
+
+        val result = createCategoryRemote(category.copy(id = id))
+        if (result.isSuccessful)
+            setSynchronized(category)
     }
 
     override suspend fun updateCategory(category: Category) {
         val local = mapper.map(category)
         localSource.updateCategory(local)
+
+        if (!tokenManager.isAuthorized()) return
+
+        val result = updateCategoryRemote(category)
+        if (result.isSuccessful)
+            setSynchronized(category)
     }
 
     override suspend fun deleteCategory(category: Category) {
         val local = mapper.map(category)
-        localSource.deleteCategory(local)
+        setMustBeDeleted(category)
+
+        if (!tokenManager.isAuthorized()) {
+            localSource.deleteCategory(local)
+            return
+        }
+
+        val result = deleteCategoryRemote(category.id)
+        if (result.isSuccessful)
+            localSource.deleteCategory(local)
     }
 
     override suspend fun getCategories(): List<Category> {
@@ -38,20 +59,34 @@ class CategoryRepositoryImpl @Inject constructor(
         return list.map { mapper.map(it) }
     }
 
-    suspend fun createCategoryRemote(category: Category): BasicResult {
+    private suspend fun createCategoryRemote(category: Category): BasicResult {
         val model = mapper.mapToCreatingModel(category)
         val response = remoteSource.createCategory(model)
         return resultMapper.map(response)
     }
 
-    suspend fun updateCategoryRemote(category: Category): BasicResult {
+    private suspend fun updateCategoryRemote(category: Category): BasicResult {
         val model = mapper.mapToEditingModel(category)
         val response = remoteSource.updateCategory(model)
         return resultMapper.map(response)
     }
 
-    suspend fun deleteCategoryRemote(id: Long): BasicResult {
+    private suspend fun deleteCategoryRemote(id: Long): BasicResult {
         val response = remoteSource.deleteCategoryById(id)
         return resultMapper.map(response)
+    }
+
+    private suspend fun setSynchronized(category: Category) {
+        val local = mapper.map(category)
+        localSource.updateCategory(
+            local.copy(isSynchronized = true)
+        )
+    }
+
+    private suspend fun setMustBeDeleted(category: Category) {
+        val local = mapper.map(category)
+        localSource.updateCategory(
+            local.copy(mustBeDeleted = true)
+        )
     }
 }
